JARVIS_MARK_21_Beta/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
auto-py-to-exe==2.46.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
bottle==0.13.3
bottle-websocket==0.2.9
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clipman==3.3.1
colorama==0.4.6
comm==0.2.2
comtypes==1.4.10
contourpy==1.2.0
cryptography==44.0.2
curl_cffi==0.10.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dateparser==1.2.1
dbus-next==0.2.3
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
docstring-inheritance==2.2.2
docutils==0.21.2
edge-tts==6.1.11
Eel==0.18.1
elevenlabs==1.57.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
fonttools==4.50.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.2
future==1.0.0
g4f==0.5.1.2
generativeai==0.0.1
gevent==25.4.1
gevent-websocket==0.10.1
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
GoogleBard1==2.1.4
googlesearch-python==1.2.4
greenlet==3.2.1
groq==0.5.0
grpcio==1.71.0
grpcio-status==1.71.0
gTTS==2.5.4
h11==0.14.0
h2==4.2.0
halo==0.0.31
Helpingai-T2==0.5
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.1
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
kivymd==1.2.0
kiwisolver==1.4.5
log-symbols==0.0.14
loguru==0.7.3
lxml==5.3.2
markdown-it-py==2.2.0
markdownify==1.1.0
MarkupSafe==3.0.2
matplotlib==3.8.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
moviepy==2.1.2
multidict==6.4.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
NetHyTech-STT==1.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
opencv-python==4.9.0.80
orjson==3.10.16
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.3.6
playsound==1.2.2
poe-api-wrapper==1.7.0
proglog==0.1.11
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pvporcupine==1.9.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.1
pydantic_core==2.33.0
pygame==2.5.2
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pyparsing==3.1.2
pypiwin32==223
pyreqwest_impersonate==0.5.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pyttsx3==2.98
pytube==15.0.0
pytz==2025.2
pywin32==308
pywin32-ctypes==0.2.3
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
quickjs==1.19.4
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.4
rpds-py==0.23.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.31.0
Send2Trash==1.8.3
setuptools==78.1.0
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.14.2
spinners==0.0.24
srt==3.5.3
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.2
termcolor==3.0.1
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tldextract==5.2.0
tls-client==1.0.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
vosk==0.3.45
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
webscout==2.9
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.13
win32_setctime==1.2.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.20.0
zope.event==5.0
zope.interface==7.2
